Add climber subystem with dummy RobotMap values and related methods.
I added a climber subsystem with some simple methods such as a way
to check if the motor is stalling or not (assuming there is a similar
way of climbing until stalling as last year). To accompany this I
added some RobotMap values as well that are dummy right now.

I made sure to have some methods update SmartDashboard / Shuffleboard
as well.