Make climber PID optional and split Climb command into three classes.
The Climb command now will climb until stall, climb until encoder distance
(PID using or not), and climb for a certain time (or indefinitely). The reason
this was split up was because with multiple constructors asking for a
direction and double time or double distance would create too much
confusion about which command was being used.

I also fleshed out the PID in the subsystem and added a lot of information
about it in the SmartDashboard and Shuffleboard view.