Add experimental mid auto code and implement dummy commands.
I implemented all of the related commands and logic per this document
https://docs.google.com/document/d/1vjeEYJNdjQMlNJY2zu3ZlKfzT9le9Da-6b9200NpEFo/
for the Mid Auto strategy. In doing this I had to add dummy Commands in DriveAngleAssisted.

I also used a lot of methods in MidAuto because a lot of code was repeated
and it was cleaner to use method calls in the CommandGroup constructor and
leave ugly logic in the methods.

All of the measurements are theoretical and were made based off of this document
outlining the different https://firstfrc.blob.core.windows.net/frc2018/Drawings/LayoutandMarkingDiagram.pdf
distances between game structures. I tried to move to distances between large
objects as much as possible.