Fix dropCubeMiddleToSwitch and other MidAuto changes.
One good TODO that I haven't done is to map on paper all of the paths that are
programmed in to check the measurements from the layouts diagram that I made.
I always tried to go halfway for wide game field elements.

Here's the link to the map I use.https://firstfrc.blob.core.windows.net/frc2018/Drawings/LayoutandMarkingDiagram.pdf